pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "israel-nadlan-data-collector"
version = "0.2.0"
description = "Public-safe collector for Israeli locality reference data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pandas>=2.2",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "rich>=13.7",
    "tenacity>=8.3",
]

[project.optional-dependencies]
dev = [
    "coverage[toml]>=7",
    "mypy",
    "pre-commit>=3.8",
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-httpx>=0.30",
    "ruff",
    "types-requests",
]

[project.scripts]
indc = "rent_collector.cli:main"
rent-collector = "rent_collector.cli:main"
rent-collect = "rent_collector.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: higher-level tests that exercise multiple modules together",
]
addopts = [
    "--strict-markers",
    "--cov=src/rent_collector",
    "--cov-report=term-missing",
]

[tool.coverage.run]
branch = true
relative_files = true
source = ["src/rent_collector"]

[tool.coverage.paths]
rent_collector = [
    "src/rent_collector",
]
